Move stl_loder and assimp_loader to own folder.
