+++ rviz_rendering/src/rviz_rendering/mesh_loader.cpp
@@ -73,8 +73,8 @@
 
 #include "resource_retriever/retriever.h"
 
-#include "assimp_loader.hpp"
-#include "stl_loader.hpp"
+#include "src/rviz_rendering/mesh_loader_helpers/assimp_loader.hpp"
+#include "src/rviz_rendering/mesh_loader_helpers/stl_loader.hpp"
 #include "rviz_rendering/logging.hpp"
 
 #define ROS_PACKAGE_NAME "rviz_rendering"
