+++ rviz_rendering/src/rviz_rendering/mesh_loader_helpers/assimp_loader.cpp
