+++ rviz_rendering/src/rviz_rendering/mesh_loader_helpers/stl_loader.cpp
